am62px: fvid2: fix STRONG.TYPE.ASSIGN.RETURN
The STRONG.TYPE.ASSIGN.RETURN checker finds function return statements
in which a strong type is expected, but not present.

Returned value is explicitly cast to the strong type to resolve the issue.

Fixes: SITSW-3530

+++ source/drivers/fvid2/v0/fvid2_utils.c
@@ -203,7 +203,7 @@ int32_t Fvid2Utils_memcmp(const void *mem1,
     GT_assert(Fvid2Trace, (NULL_PTR != mem1));
     GT_assert(Fvid2Trace, (NULL_PTR != mem2));
 
-    return memcmp(mem1, mem2, byteCount);
+    return (int32_t)(memcmp(mem1, mem2, byteCount));
 }
 
 /**
